requirements.txt
python-telegram-bot==21.0.1
gspread==6.0.2
google-auth==2.27.0
